pyproject.toml
[tool.poetry]
name = "turbinia"
version = "20240930"
description = "Automation and Scaling of Digital Forensics Tools"
authors = ["Turbinia Developers <turbinia-dev@googlegroups.com>"]
maintainers = ["Turbinia Developers <turbinia-dev@googlegroups.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/google/turbinia"
documentation = "https://turbinia.readthedocs.io/en/latest/"

[tool.poetry.scripts]
turbiniactl = "turbinia.turbiniactl:main"

[tool.poetry.dependencies]
python = "^3.10"
acstore = { version = "20240128" }
backoff = { version = ">=2.2.1" }
celery = { version = "^5.2.2" }
dfDewey = { version = "^20231016", optional = true }
dfimagetools = { version = "^20240301", optional = true }
docker = { version = "^6.1.3" }
fastapi = {extras = ["all"], version = ">=0.75.0,<0.99.0"}
filelock = { version = "*" }
google-api-core = { version = "<3.0.0", optional = true }
google-generativeai = { version = ">=0.8.2" }
libcloudforensics = "^20240325"
lief = { version = ">= 0.15.1", optional = true }
opensearch-py = {version = "2.4.2"}
pandas = { version = "^2.1.0" }
plaso = { version = "20240308", optional = true }
pyssdeep = { version = "1.0.0", optional = true }
prometheus_client = { version = "^0.17.1" }
protobuf = { version = ">=3.19.0", optional = true }
pydantic = { version = "^1.10.5,<2"}
pyglove = { version = ">=0.4.4" }
pyhindsight = { version = "^20230327.0", optional = true }
py-tlsh = { version = ">= 4.7.2", optional = true }
ratelimit = { version = ">=2.2.1" }
redis = { version = "^4.4.4" }
urllib3 = [
  {version = ">=1.25.4,<1.27", python = "<3.10"},
  {version = ">=1.25.4,<2.1", python = ">=3.10"}
]
debugpy = "^1.8.0"
jurigged = "^0.5.7"
gitpython = "^3.1.43"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "*"
fakeredis = "^1.8.1"
google-auth-oauthlib = "^1.1.0"
mock = "*"
pytest = "*"
tabulate = "*"
turbinia-api-lib = "^1.0.6"
yapf = "*"

[tool.poetry.extras]
worker = [
  "dfimagetools",
  "dfDewey",
  "lief",
  "plaso",
  "pyssdeep",
  "pyhindsight",
  "py-tlsh",
]

[tool.yapfignore]
ignore_patterns = [
  "turbinia/api/client/**/*.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
